caProvider: Convert caChannelList into a tsDLList
This should provide a significant performance boost when creating
many thousands of CA channels. The only time it is necessary to
traverse the list is in the CAChannelProvider's destructor; when
individual channels are added or destroyed they can insert or
remove themselves from the list without having to do a search.